+++ numalgsolve/Division.py
@@ -1,12 +1,13 @@
 import numpy as np
+import itertools
 from numalgsolve.utils import get_var_list, slice_top, row_swap_matrix, mon_combos, newton_polish
-from numalgsolve.TelenVanBarel import add_polys, rrqr_reduceTelenVanBarel, rrqr_reduceTelenVanBarel2
+from numalgsolve.TVBCore import add_polys, rrqr_reduceTelenVanBarel, makeBasisDict
 from scipy.linalg import solve_triangular, eig, qr
 from matplotlib import pyplot as plt
 
 def division_cheb(polys, divisor_var = 0, tol = 1.e-12):
     '''Calculates the common zeros of polynomials using a division matrix.
-    
+
     Parameters
     --------
     polys: MultiCheb Polynomials
@@ -19,7 +20,7 @@ def division_cheb(polys, divisor_var = 0, tol = 1.e-12):
     zeros : list
         The common zeros of the polynomials. Each list element is a numpy array of complex entries
         that contains the coordinates in each dimension of the zero.
-    '''    
+    '''
     #This first section creates the Macaulay Matrix with the monomials that don't have
     #the divisor variable in the first columns.
     dim = polys[0].dim
@@ -28,43 +29,45 @@ def division_cheb(polys, divisor_var = 0, tol = 1.e-12):
     poly_coeff_list = []
     for poly in polys:
         poly_coeff_list = add_polys(matrix_degree, poly, poly_coeff_list)
-    
+
     matrix, matrix_terms, cuts = create_matrix(poly_coeff_list, matrix_degree, dim, divisor_var)
-        
+
     #Reduces the Macaulay matrix like normal.
     matrix, matrix_terms = rrqr_reduceTelenVanBarel(matrix, matrix_terms, cuts, tol)
-    
+
     rows,columns = matrix.shape
     VB = matrix_terms[matrix.shape[0]:]
     matrix = np.hstack((np.eye(rows),solve_triangular(matrix[:,:rows],matrix[:,rows:])))
 
+
+#------------> chebyshev
     #Builds the inverse matrix. The terms are the vector basis as well as y^k/x terms for all k. Reducing
     #this matrix allows the y^k/x terms to be reduced back into the vector basis.
     inverses = matrix_terms[np.where(matrix_terms[:,divisor_var] == 0)[0]]
-    inverses[:,divisor_var] = -np.ones(inverses.shape[0], dtype = 'int')    
+    inverses[:,divisor_var] = -np.ones(inverses.shape[0], dtype = 'int')
     inv_matrix_terms = np.vstack((inverses, VB))
     inv_matrix = np.zeros([len(inverses),len(inv_matrix_terms)])
 
     #A bunch of different dictionaries are used below for speed purposes and to prevent repeat calculations.
-    
+
     #A dictionary of term in inv_matrix_terms to their spot in inv_matrix_terms.
     inv_spot_dict = dict()
     spot = 0
     for term in inv_matrix_terms:
         inv_spot_dict[tuple(term)] = spot
         spot+=1
-    
+
     #A dictionary of terms on the diagonal to their reduction in the vector basis.
     diag_reduction_dict = dict()
     for i in range(matrix.shape[0]):
         term = matrix_terms[i]
         diag_reduction_dict[tuple(term)] = matrix[i][-len(VB):]
-    
+
     #A dictionary of terms to the terms in their quotient when divided by x.
     divisor_terms_dict = dict()
     for term in matrix_terms:
         divisor_terms_dict[tuple(term)] = get_divisor_terms(term, divisor_var)
-        
+
     #A dictionary of terms to their quotient when divided by x.
     term_divide_dict = dict()
     for term in matrix_terms[-len(VB):]:
@@ -80,7 +83,7 @@ def division_cheb(polys, divisor_var = 0, tol = 1.e-12):
 
     #Reduces the inv_matrix to solve for the y^k/x terms in the vector basis.
     Q,R = qr(inv_matrix)
-        
+
     inv_solutions = np.hstack((np.eye(R.shape[0]),solve_triangular(R[:,:R.shape[0]], R[:,R.shape[0]:])))
 
     #A dictionary of term in the vector basis to their spot in the vector basis.
@@ -97,8 +100,30 @@ def division_cheb(polys, divisor_var = 0, tol = 1.e-12):
 
     #Builds the division matrix and finds the eigenvalues and eigenvectors.
     division_matrix = build_division_matrix(VB, VB_spot_dict, diag_reduction_dict, inv_reduction_dict, divisor_terms_dict)
+#<---------end Chebyshev
+#--------->Power
+    """basisDict = makeBasisDict(matrix, matrix_terms, VB)
+
+    #Dictionary of terms in the vector basis their spots in the matrix.
+    VBdict = {}
+    spot = 0
+    for row in VB:
+        VBdict[tuple(row)] = spot
+        spot+=1
+
+    # Build division matrix
+    division_matrix = np.zeros((len(VB), len(VB)))
+    for i in range(VB.shape[0]):
+        var = np.zeros(dim)
+        var[divisor_var] = 1
+        term = tuple(VB[i] - var)
+        if term in VBdict:
+            division_matrix[VBdict[term]][i] += 1
+        else:
+            division_matrix[:,i] -= basisDict[term]"""
+#<----------end Power
     vals, vecs = eig(division_matrix.T)
-    
+
     #Calculates the zeros, the x values from the eigenvalues and the y values from the eigenvectors.
     zeros = list()
     for i in range(len(vals)):
@@ -117,7 +142,7 @@ def division_cheb(polys, divisor_var = 0, tol = 1.e-12):
 
 def get_matrix_terms(poly_coeffs, dim, divisor_var, deg):
     '''Finds the terms in the Macaulay matrix.
-    
+
     Parameters
     --------
     poly_coeffs: list
@@ -133,7 +158,7 @@ def get_matrix_terms(poly_coeffs, dim, divisor_var, deg):
         The matrix_terms. The ith row is the term represented by the ith column of the matrix.
     cuts : tuple
         When the matrix is reduced it is split into 3 parts with restricted pivoting. These numbers indicate
-        where those cuts happen.    
+        where those cuts happen.
     '''
     """
     #The following code is just for testing, just two dimensional divide by x.
@@ -153,7 +178,7 @@ def get_matrix_terms(poly_coeffs, dim, divisor_var, deg):
                 matrix_term_set_y.add(term)
             else:
                 matrix_term_set_other.add(term)
-    
+
     needed_terms = list()
     base = np.zeros(dim, dtype = 'int')
     base[divisor_var] = 1
@@ -168,11 +193,51 @@ def get_matrix_terms(poly_coeffs, dim, divisor_var, deg):
             base[i] = 0
     for term in needed_terms:
         matrix_term_set_other.remove(term)
-    
+
     matrix_terms = np.vstack((np.vstack(matrix_term_set_y),np.vstack(matrix_term_set_other),matrix_term_end))
-        
+
     return matrix_terms, tuple([len(matrix_term_set_y), len(matrix_term_set_y)+len(matrix_term_set_other)])
-    
+
+def makeBasisDict(matrix, matrix_terms, VB):
+    '''Calculates and returns the basisDict.
+
+    This is a dictionary of the terms on the diagonal of the reduced TVB matrix to the terms in the Vector Basis.
+    It is used to create the multiplication matrix in root_finder.
+
+    Parameters
+    --------
+    matrix: numpy array
+        The reduced TVB matrix.
+    matrix_terms : numpy array
+        The terms in the matrix. The i'th row is the term represented by the i'th column of the matrix.
+    VB : numpy array
+        Each row is a term in the vector basis.
+
+    Returns
+    -----------
+    basisDict : dict
+        Maps terms on the diagonal of the reduced TVB matrix (tuples) to numpy arrays of the shape remainder_shape
+        that represent the terms reduction into the Vector Basis.
+    '''
+    basisDict = {}
+
+    VBSet = set()
+    for i in VB:
+        VBSet.add(tuple(i))
+
+    #We don't actually need most of the rows, so we only get the ones we need.
+    neededSpots = set()
+    for term, mon in itertools.product(VB,get_var_list(VB.shape[1])):
+        if tuple(term-mon) not in VBSet:
+            neededSpots.add(tuple(term-mon))
+
+    for i in range(matrix.shape[0]):
+        term = tuple(matrix_terms[i])
+        if term not in neededSpots:
+            continue
+        basisDict[term] = matrix[i][matrix.shape[0]:]
+
+    return basisDict
 
 def create_matrix(poly_coeffs, degree, dim, divisor_var):
     ''' Builds a Macaulay matrix for reduction.
@@ -187,7 +252,7 @@ def create_matrix(poly_coeffs, degree, dim, divisor_var):
         The dimension of the polynomials going into the matrix.
     divisor_var : int
         What variable is being divided by. 0 is x, 1 is y, etc. Defaults to x.
-    
+
     Returns
     -------
     matrix : 2D numpy array
@@ -226,7 +291,7 @@ def create_matrix(poly_coeffs, degree, dim, divisor_var):
 
 def divide_row(coeffs, terms, term_divide_dict, length):
     """Divides a row of the matrix by the divisor variable..
-    
+
     Parameters
     ----------
     coeffs : numpy array.
@@ -241,7 +306,7 @@ def divide_row(coeffs, terms, term_divide_dict, length):
     -------
     new_row : numpy array
         The row we get in the inverse_matrix by dividing the first row by x.
-    """    
+    """
     new_row = np.zeros(length)
     for i in range(len(coeffs)):
         new_row+=coeffs[i]*term_divide_dict[tuple(terms[i])]
@@ -249,7 +314,7 @@ def divide_row(coeffs, terms, term_divide_dict, length):
 
 def divide_term(term, inv_matrix_terms, inv_spot_dict, diag_reduction_dict, VB_size, divisor_terms_dict):
     """Divides a term of the matrix by the divisor variable.
-    
+
     Parameters
     ----------
     term: numpy array
@@ -264,7 +329,7 @@ def divide_term(term, inv_matrix_terms, inv_spot_dict, diag_reduction_dict, VB_s
         The number of elements in the vector basis.
     divisor_terms_dict : dictionary
         A dictionary of terms to the terms in their dividend when divided by x.
-    
+
     Returns
     -------
     row : numpy array
@@ -288,12 +353,12 @@ def divide_term(term, inv_matrix_terms, inv_spot_dict, diag_reduction_dict, VB_s
 
 def get_divisor_terms(term, divisor_var):
     """Finds the terms that will be present when dividing a given term by x.
-    
+
     Parameters
     ----------
     term: numpy array
         The term to divide.
-    
+
     Returns
     -------
     terms : numpy array
@@ -315,7 +380,7 @@ def get_divisor_terms(term, divisor_var):
 
 def build_division_matrix(VB, VB_spot_dict, diag_reduction_dict, inv_reduction_dict, divisor_terms_dict):
     """Builds the division matrix.
-    
+
     Parameters
     ----------
     VB: numpy array
@@ -328,7 +393,7 @@ def build_division_matrix(VB, VB_spot_dict, diag_reduction_dict, inv_reduction_d
         A dictionary of terms of type y^k/x to their reduction in the vector basis.
     divisor_terms_dict : dictionary
         A dictionary of terms to the terms in their dividend when divided by x.
-    
+
     Returns
     -------
     row : numpy array
