+++ numalgsolve/subdivision.py
@@ -9,7 +9,7 @@ the approximation degree is small enough to be solved efficiently.
 import numpy as np
 from numpy.fft.fftpack import fftn
 from numalgsolve.OneDimension import divCheb,divPower,multCheb,multPower,solve
-from numalgsolve.ChebyshevDivision import division_cheb
+from numalgsolve.Division import division_cheb
 from numalgsolve.utils import clean_zeros_from_matrix, slice_top
 from numalgsolve.polynomial import MultiCheb
 
@@ -61,7 +61,7 @@ def solve(funcs, a, b):
 
 def transform(x,a,b):
     """Transforms points from the interval [-1,1] to the interval [a,b].
-    
+
     Parameters
     ----------
     x : numpy array
@@ -70,7 +70,7 @@ def transform(x,a,b):
         The lower bound on the interval. Float if one-dimensional, numpy array if multi-dimensional
     b : float or numpy array
         The upper bound on the interval. Float if one-dimensional, numpy array if multi-dimensional
-    
+
     Returns
     -------
     transform : numpy array
@@ -80,7 +80,7 @@ def transform(x,a,b):
 
 def inv_transform(x,a,b):
     """Transforms points from the interval [a,b] to the interval [-1,1].
-    
+
     Parameters
     ----------
     x : numpy array
@@ -89,7 +89,7 @@ def inv_transform(x,a,b):
         The lower bound on the interval. Float if one-dimensional, numpy array if multi-dimensional
     b : float or numpy array
         The upper bound on the interval. Float if one-dimensional, numpy array if multi-dimensional
-    
+
     Returns
     -------
     transform : numpy array
@@ -99,7 +99,7 @@ def inv_transform(x,a,b):
 
 def interval_approximate_1d(f,a,b,deg):
     """Finds the chebyshev approximation of a one-dimensional function on an interval.
-    
+
     Parameters
     ----------
     f : function from R -> R
@@ -110,7 +110,7 @@ def interval_approximate_1d(f,a,b,deg):
         The upper bound on the interval.
     deg : int
         The degree of the interpolation.
-    
+
     Returns
     -------
     coeffs : numpy array
@@ -125,7 +125,7 @@ def interval_approximate_1d(f,a,b,deg):
 
 def interval_approximate_nd(f,a,b,degs):
     """Finds the chebyshev approximation of an n-dimensional function on an interval.
-    
+
     Parameters
     ----------
     f : function from R^n -> R
@@ -136,7 +136,7 @@ def interval_approximate_nd(f,a,b,degs):
         The upper bound on the interval.
     deg : numpy array
         The degree of the interpolation in each dimension.
-    
+
     Returns
     -------
     coeffs : numpy array
@@ -181,7 +181,7 @@ def interval_approximate_nd(f,a,b,degs):
             for spot,val in np.ndenumerate(values):
                 values[spot] = vals[i]
                 i+=1
-    
+
     coeffs = np.real(fftn(values/np.product(degs)))
 
     for i in range(dim):
@@ -200,7 +200,7 @@ def interval_approximate_nd(f,a,b,degs):
 
 def get_subintervals(a,b,dimensions):
     """Gets the subintervals to divide a matrix into.
-    
+
     Parameters
     ----------
     a : numpy array
@@ -209,7 +209,7 @@ def get_subintervals(a,b,dimensions):
         The upper bound on the interval.
     dimensions : numpy array
         The dimensions we want to cut in half.
-    
+
     Returns
     -------
     subintervals : list
@@ -230,9 +230,9 @@ def get_subintervals(a,b,dimensions):
 
 def interval_perms(perms,perm,spot):
     """A helper function for get_subintervals. Finds the different ways we can split the dimensions.
-    
+
     Called recursively.
-    
+
     Parameters
     ----------
     perms : list
@@ -241,7 +241,7 @@ def interval_perms(perms,perm,spot):
         The way we are currently dynamically computing.
     spot : int
         Where in perm we are changing.
-    
+
     Returns
     -------
     subintervals : list
@@ -260,9 +260,9 @@ def interval_perms(perms,perm,spot):
 
 def full_cheb_approximate(f,a,b,deg,tol=1.e-8):
     """Gives the full chebyshev approximation and checks if it's good enough.
-    
+
     Called recursively.
-    
+
     Parameters
     ----------
     f : function
@@ -275,14 +275,14 @@ def full_cheb_approximate(f,a,b,deg,tol=1.e-8):
         The degree to approximate with.
     tol : float
         How small the high degree terms must be to consider the approximation accurate.
-    
+
     Returns
     -------
     coeff : numpy array
         The coefficient array of the interpolation. If it can't get a good approximation and needs to subdivide, returns None.
     """
     dim = len(a)
-    
+
     degs = np.array([deg]*dim)+1
     coeff = interval_approximate_nd(f,a,b,degs)
     coeff2 = interval_approximate_nd(f,a,b,degs*2)
@@ -297,14 +297,14 @@ def full_cheb_approximate(f,a,b,deg,tol=1.e-8):
 
 def good_zeros_nd(zeros, imag_tol = 1.e-10):
     """Get the real zeros in the -1 to 1 interval in each dimension.
-    
+
     Parameters
     ----------
     zeros : numpy array
         The zeros to be checked.
     imag_tol : float
         How large the imaginary part can be to still have it be considered real.
-    
+
     Returns
     -------
     good_zeros : numpy array
@@ -317,7 +317,7 @@ def good_zeros_nd(zeros, imag_tol = 1.e-10):
 
 def subdivision_solve_nd(funcs,a,b,deg,tol=1.e-8,tol2=1.e-8):
     """Finds the common zeros of the given functions.
-    
+
     Parameters
     ----------
     funcs : list
@@ -328,7 +328,7 @@ def subdivision_solve_nd(funcs,a,b,deg,tol=1.e-8,tol2=1.e-8):
         The upper bound on the interval.
     deg : int
         The degree to approximate with in the chebyshev approximation.
-    
+
     Returns
     -------
     good_zeros : numpy array
@@ -354,7 +354,7 @@ def subdivision_solve_nd(funcs,a,b,deg,tol=1.e-8,tol2=1.e-8):
 
 def trim_coeff(coeff, tol=1.e-8, tol2=1.e-8):
     """Finds the common zeros of the given functions.
-    
+
     Parameters
     ----------
     funcs : list
@@ -365,7 +365,7 @@ def trim_coeff(coeff, tol=1.e-8, tol2=1.e-8):
         The upper bound on the interval.
     deg : int
         The degree to approximate with in the chebyshev approximation.
-    
+
     Returns
     -------
     good_zeros : numpy array
@@ -378,9 +378,9 @@ def trim_coeff(coeff, tol=1.e-8, tol2=1.e-8):
         for s in spot:
             slices.append(s)
         coeff[slices] = 0
-    
+
     dim = len(coeff.shape)
-    
+
     #Cuts out the high diagonals as much as possible to minimize polynomial degree.
     if abs(coeff[tuple([-1]*dim)]) < tol:
         coeff[tuple([-1]*dim)] = 0
@@ -390,13 +390,13 @@ def trim_coeff(coeff, tol=1.e-8, tol2=1.e-8):
             slices = list()
             mons = np.array(mons).T
             for i in range(dim):
-                slices.append(mons[i])        
+                slices.append(mons[i])
             if np.sum(np.abs(coeff[slices])) < tol:
                 coeff[slices] = 0
             else:
                 break
             deg -= 1
-    
+
     return coeff
 
 def mon_combos_limited(mon, numLeft, shape, spot = 0):
@@ -439,14 +439,14 @@ def mon_combos_limited(mon, numLeft, shape, spot = 0):
 
 def good_zeros(zeros, imag_tol = 1.e-10):
     """Get the real zeros in the -1 to 1 interval
-    
+
     Parameters
     ----------
     zeros : numpy array
         The zeros to be checked.
     imag_tol : float
         How large the imaginary part can be to still have it be considered real.
-    
+
     Returns
     -------
     good_zeros : numpy array
@@ -458,7 +458,7 @@ def good_zeros(zeros, imag_tol = 1.e-10):
 
 def subdivide_solve_1d(f,a,b,cheb_approx_tol=1.e-10,max_degree=128):
     """Finds the roots of a one-dimensional function using subdivision and chebyshev approximation.
-    
+
     Parameters
     ----------
     f : function from R^n -> R
@@ -469,7 +469,7 @@ def subdivide_solve_1d(f,a,b,cheb_approx_tol=1.e-10,max_degree=128):
         The upper bound on the interval.
     deg : int
         The degree of the interpolation.
-    
+
     Returns
     -------
     coeffs : numpy array
